Split operators into separate files
