+++ loica/__init__.py
@@ -2,6 +2,6 @@ from .assay import *
 from .sample import *
 from .genetic_network import *
 from .metabolism import *
-from .operator import *
+from .operators import *
 from .geneproduct import *
 from .supplement import *
