+++ loica/operators/not_.py
@@ -1,0 +1,16 @@
+class Not:
+    def __init__(self, input, output, a, b, K, n):
+        self.a = a
+        self.b = b
+        self.K = K
+        self.n = n
+        self.input = input
+        self.output = output
+        
+    def expression_rate(self):
+        input_repressor = self.input.concentration
+        r = (input_repressor/self.K)**self.n
+        expression_rate = ( self.a + self.b*r ) / (1 + r)
+        return expression_rate
+
+
